Enhance FTP Server's initialization and logging
This commit modifies the 'New' function in FTP server configuration to now take an address as an argument. This affords a higher degree of flexibility when setting up servers. Moreover, logging levels for client connection and disconnection events have been raised from Debug to Info. Additionally, the process will register more details, such as the client version. A fail-safe mechanism has also been introduced - If neither FTPAddr nor HTTPAddr is specified, the FTP server will default to a preset address. This set of changes aims to bolster the server's reliability and traceability.